pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Bagpyp <robert@bagpyp.net>"]
readme = "README.md"
packages = [
    { include = "src", from = "."},
]
package-mode = true


[tool.poetry.dependencies]
python = "^3.13"
black = {extras = ["jupyter"], version = "^25.1.0"}
python-dotenv = "^1.0.1"
alembic = "^1.15.1"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
